// frontend/src/shared/lib/config/preferences-registry.d.ts
export interface PreferenceLocale {
  language: string;
  timezone: string;
  date_format: "YYYY-MM-DD" | "MM/DD/YYYY" | "DD/MM/YYYY" | "DD-MM-YYYY" | "MMM D, YYYY" | "D MMM YYYY" | "MMMM D, YYYY" | "D MMMM YYYY" | "ddd, MMM D, YYYY" | "MM.DD.YYYY" | "DD.MM.YYYY";
  time_format: "12h" | "24h";
}